+++ asm/custom-functions/src/game/mod.rs
@@ -1,0 +1,13 @@
+pub mod actor;
+pub mod arc;
+pub mod bird;
+pub mod events;
+pub mod file_manager;
+pub mod flag_managers;
+pub mod item;
+pub mod message;
+pub mod minigame;
+pub mod player;
+pub mod reloader;
+pub mod save_file;
+pub mod stage_info;