+++ logic/logic_utils.py
@@ -27,6 +27,7 @@ class AdditionalInfo:
     randomized_dungeon_entrance: dict[str, str]
     randomized_trial_entrance: dict[str, str]
     randomized_start_entrance: dict[str, str]
+    randomized_start_statues: dict[str, str]
     known_locations: List[EIN]
 
 
@@ -54,6 +55,7 @@ class LogicUtils(Logic):
         self.randomized_dungeon_entrance = additional_info.randomized_dungeon_entrance
         self.randomized_trial_entrance = additional_info.randomized_trial_entrance
         self.randomized_start_entrance = additional_info.randomized_start_entrance
+        self.randomized_start_statues = additional_info.randomized_start_statues
         self.known_locations = additional_info.known_locations
 
     def check(self, useroutput):